a try for return an array from compute_chi_2_conditional_test_indep

+++ smmb_aco/include/statistics/statistics.hpp
@@ -22,7 +22,7 @@ public:
     // associated score and p value
     //==========================================================================
     static float make_contingencies_chi_2_conditional_test_indep(boost::numeric::ublas::matrix_column<boost::numeric::ublas::matrix<int>> const& _genos_column, boost_vector_int const& _phenos_vector, std::list<unsigned> const& cond_genos_indexes);
-    static float compute_chi_2_conditional_test_indep(std::vector<contingencies> contingencies_vector, unsigned int liberty_degree, unsigned int number_obs_subset);
+    static int compute_chi_2_conditional_test_indep(std::vector<contingencies> contingencies_vector, unsigned int liberty_degree, unsigned int number_obs_subset);
     //==========================================================================
     // Tools for statitics
     //==========================================================================
